Initial 9DoF sensor calculations
This is not integrated at all with the rest of the /main files. This is a preliminary idea as to how the 9DoF sensor may be implemented in the future. Deviation calculations use the gyroscope's X and Z axes (and if deviations in rotational velocity occur), which may be the wrong way to calculate if the rocket deviates.

Uses Arduino's Serial.print to display data in the Serial Monitor (usually located in the Arduino IDE; not sure if Linux is able to view this).

The programs should theoretically work, but no testing has been done.

Requires the SparkFunLSM9DS1.h library (located https://github.com/sparkfun/SparkFun_LSM9DS1_Arduino_Library).